+++ routes/api/BixiStation.js
@@ -1,8 +1,6 @@
-const express = require('express');
+const router = require('express').Router();
+const bixiStationsController = require("../../controllers/bixiStationsController");
 const db = require('../../models');
-const getRealTimeData = require('../../utils/api/StationsData');
-
-const router = express.Router();
 
 // Calculate distance in m between two coordinates  
 // const calculateDistance = (location1, location2) => {
@@ -22,38 +20,14 @@ const router = express.Router();
 // Return the closest 10 BixiStations, based on searched location
 // Note: coord_diff correlates to distance well enough - Google used on
 // front end to calcualte actual path distance for each plotted object
-router.get('/bixi/', (req, res) => {
-
-    db.BixiStation.findAll({
-        attributes:['id', 'idx', 'name', 'lat', 'lng', 'capacity', 
-            [db.sequelize.literal(`ABS(lat - ${req.query.lat}) + ABS(lng - ${req.query.lng})`), 'coord_diff']],
-        order: db.sequelize.col('coord_diff'),
-        limit: 10
-    })
-    .then(data => {
-        res.status(200).send(data);
-    })
-    .catch(err => {
-        res.status(400).end();
-    })
-});
+router.route('/')
+    .get(bixiStationsController.getNearest);
 
 // Return real time station data for stations identified by idx
-router.get('/bixi/realtime', (req, res) => {
-        getRealTimeData((data) => {
-            const stationsArray = [];
+router.route('/realtime')
+    .get(bixiStationsController.getRealTime);
 
-            try{            
-                req.body.idxArray.forEach(idx => {
-                    stationsArray.push(data[idx]);
-                });
-
-                res.status(200).send(stationsArray);
-            }
-            catch{
-                res.status(400).end();
-            }
-        });
-});
+router.route('/seed')
+    .get(bixiStationsController.seed);
 
 module.exports = router;
